Beginning to add GPU support to oat_gcp.sh
The new "enable_gpu" variable in `oat_gcp_config` is used to determine
whether we need to install the NVIDIA driver and the
nvidia-container-toolkit package in the VM. The instance creation
command is also a bit different when requesting a GPU.

The other change here is to use `sed` to remove the sections of
`docker-compose.yml` which declare a GPU is required for certain
services. Since this is enabled by default it will cause a Docker error
when trying to launch `llm_functionalities` on an instance without a
GPU, leaving the system in a partially online state. By removing the
"deploy:" sections in the service configurations it allows
`llm_functionalities` to start and then exit gracefully because the code
inside it checks for the presence of a GPU.