Initial attempt to integrate TGI
I've replaced the existing code to load the model with creating an
`InferenceClient` object using the endpoint URL defined in the
docker-compose.yml file.

Creating the object doesn't trigger any connection, so it currently
submits a simple query to check if the TGI endpoint is actually
available. This might not happen immediately (e.g. if it's still
downloading or loading a model), it currently has a basic retry setup
but might need some more thought put into it.

The `call_model` and `batch_call_model` methods are updated to call the
`.text_generation` method on the `InferenceClient`. For the batch
method, it submits things in parallel using a ThreadPoolExecutor, TGI
doesn't offer a batch-specific endpoint but it should automatically batch
the requests internally based on the docs.