Fix parsing of image_url_lookup.json
The original version of this file seems to have been a list of
single-entry dicts, i.e.:
```
[
    {file1: url1},
    {file2: url2},
    ...
]
```
so it needed some logic in the augmenters which use it to be parsed
back to a single larger dict.

However after the OAT v2.0 updates, the current version of the file is
already a single large dict. This causes an exception when the
augmenters try to load it.

Since the file is already in the required format I've just changed the
code to load and use it directly.